연결 테스트

+++ danusys-drone/src/main/java/com/danusys/web/drone/dto/request/DroneDetailRequest.java
@@ -14,6 +14,6 @@ public class DroneDetailRequest {
     private long droneBase;
     private long droneMission;
     private String droneStatus;
-    private int droneSocket;
+    private Long droneSocket;
 
 }
